pyproject.toml
[project]
name = "ai-assistant"
version = "0.1.0"
description = "A personal AI assistant with Streamlit frontend, Google Auth, calendar integration, and knowledge base"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    
    # Streamlit Frontend
    "streamlit>=1.30.0",
    "streamlit-authenticator>=0.3.0",
    
    # Google OAuth & Calendar Integration
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth>=2.25.0",
    
    # Slack Integration
    "slack-sdk>=3.27.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    
    # Scheduling
    "schedule>=1.2.0",
    "apscheduler>=3.10.0",
    
    # Async support
    "aiohttp>=3.9.0",
    
    # Data handling
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    
    # Utilities
    "python-dateutil>=2.8.0",
    "pytz>=2024.1",
    "markdown>=3.5.0",
]

[project.scripts]
scheduler = "src.scheduler:run_scheduler"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
